Isolate the usage of the pystac-client library only for AuxipClient, CadipClient and CatalogClient
